BLOCKS-048 moved block definition files under ./js/ folder [AK]
