+++ src/js/blocks/loader.js
@@ -1,0 +1,17 @@
+import "./colours";
+import "./default_toolbox";
+
+import "./categories/arduino";
+import "./categories/control";
+import "./categories/data";
+import "./categories/drone";
+import "./categories/event";
+import "./categories/jumpingSumo";
+import "./categories/legoEV3";
+import "./categories/legoNXT";
+import "./categories/looks";
+import "./categories/motion";
+import "./categories/pen";
+import "./categories/phiro";
+import "./categories/raspi";
+import "./categories/sound";
