--- src/js/playground/toolbox/loader.js
@@ -1,17 +1,0 @@
-import "./colours";
-import "./default_toolbox";
-
-import "./blocks/arduino";
-import "./blocks/control";
-import "./blocks/data";
-import "./blocks/drone";
-import "./blocks/event";
-import "./blocks/jumpingSumo";
-import "./blocks/legoEV3";
-import "./blocks/legoNXT";
-import "./blocks/looks";
-import "./blocks/motion";
-import "./blocks/pen";
-import "./blocks/phiro";
-import "./blocks/raspi";
-import "./blocks/sound";
